@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "idtrack"
-version = "0.0.4"  # <<COOKIETEMPLE_FORCE_BUMP>>
+version = "0.0.4"
 description = "Cross-Temporal and Cross-Database Biological Identifier Mapping."
 authors = ["Kemal Inecik <k.inecik@gmail.com>"]
 license = "BSD-3-Clause"
@@ -18,28 +18,30 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: BSD License",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Operating System :: MacOS :: MacOS X",
     "Operating System :: Microsoft :: Windows",
     "Operating System :: POSIX :: Linux",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
 
 
 [tool.poetry.dependencies]
-python = ">=3.8,<4.0"
+python = ">=3.9,<4.0"
 # dependencies are versioned based on their version on 19 October 2019, when Python 3.8 was released.
 PyYAML = ">=5.2"
 h5py = ">=2.10.0"
-tables = ">=2.6.1"
-pandas = ">=0.25.3,<2.0.0"
+pandas = ">=0.25.3"
 numpy = ">=1.17.4"
 networkx = ">=2.4"
 requests = ">=2.22.0"
 PyMySQL = ">=0.9.3"
 tqdm = ">=4.37.0"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 click = ">=8.0.0"
 rich = ">=10.3.0"
 pytest = ">=6.2.3"
@@ -62,7 +64,8 @@ pre-commit-hooks = ">=4.0.1"
 sphinx-rtd-theme = ">=0.5.0"
 sphinx-click = ">=3.0.0"
 Pygments = ">=2.8.1"
-types-pkg-resources = ">=0.1.2"
+# types-pkg-resources, which was unceremoniously disappeared from the Python supply chain. instead `types-setuptools`
+types-setuptools = ">=71.1.0"
 types-requests = ">=2.25.2"
 types-attrs = ">=19.1.0"
 sphinx-rtd-dark-mode = ">=1.2.3"
@@ -83,6 +86,32 @@ idtrack = "idtrack.__main__:main"
 
 [tool.black]
 line-length = 120
+
+[tool.isort]
+known_third_party = ["GPUtil", "h5py", "networkx", "nox", "numpy", "pandas", "psutil", "pymysql", "pytest", "requests", "rich", "torch", "tqdm", "yaml"]
+multi_line_output = 3
+include_trailing_comma = true
+balanced_wrapping = true
+line_length = 120
+profile = "black"
+
+[tool.bandit]
+# (optional) list included tests here:
+tests = []
+# (optional) list skipped tests here:
+skips = ["B403", "B404", "B603", "B607"]
+
+[tool.flake8]
+select = ["B", "B9", "C", "D", "DAR", "E", "F", "N", "RST", "S", "W"]
+ignore = ["E203", "E501", "RST201", "RST203", "RST301", "W503", "D100", "B905"]
+max-line-length = 120
+max-complexity = 50
+docstring-convention = "google"
+per-file-ignores = ["tests/*:S101"]
+rst-roles = ["py:attr", "py:meth", "py:class"]
+
+[tool.darglint]
+strictness = "short"
 
 [tool.mypy]
 strict = false
