@@ -40,14 +40,14 @@ default_dark_mode = False  #
 templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
-source_suffix = ".rst"
+source_suffix = {".rst": "restructuredtext"}
 
 # The master toctree document.
 master_doc = "index"
 
 # General information about the project.
 project = "idtrack"
-copyright = "2022, Kemal Inecik"
+copyright = "2025, Kemal Inecik"
 author = "Kemal Inecik"
 
 # The version info for the project you're documenting, acts as replacement
