@@ -2,8 +2,6 @@
 
 # Kemal Inecik
 # k.inecik@gmail.com
-
-
 """Top-level package for idtrack."""
 
 __author__ = "Kemal Inecik"
@@ -11,10 +9,10 @@ __email__ = "k.inecik@gmail.com"
 __version__ = "0.0.4"
 
 
-from ._api import API  # noqa
-from ._database_manager import DatabaseManager  # noqa
-from ._db import DB  # noqa
-from ._external_databases import ExternalDatabases  # noqa
-from ._the_graph import TheGraph  # noqa
-from ._track import Track  # noqa
-from ._verify_organism import VerifyOrganism  # noqa
+from idtrack._api import API  # noqa
+from idtrack._database_manager import DatabaseManager  # noqa
+from idtrack._db import DB  # noqa
+from idtrack._external_databases import ExternalDatabases  # noqa
+from idtrack._the_graph import TheGraph  # noqa
+from idtrack._track import Track  # noqa
+from idtrack._verify_organism import VerifyOrganism  # noqa
